requirements.txt
PyPDF2
python-dotenv
requests
sentence_transformers
torch